requirements.txt
api==0.0.7
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
idna==3.2
mysqlclient==2.0.3
nose==1.3.7
pytz==2021.1
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.6
